requirements.txt
altgraph==0.17.3
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
future==0.18.3
idna==3.4
maxminddb==2.4.0
maxminddb-geolite2==2018.703
pefile==2023.2.7
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
python-whois==0.8.0
pywin32-ctypes==0.2.2
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.4
